+++ flexmock/exceptions.py
@@ -1,0 +1,33 @@
+"""Flecmock exceptions."""
+
+
+class FlexmockError(Exception):
+    """FlexmockError"""
+
+
+class MockBuiltinError(Exception):
+    """MockBuiltinError"""
+
+
+class MethodSignatureError(FlexmockError):
+    """MethodSignatureError"""
+
+
+class ExceptionClassError(FlexmockError):
+    """ExceptionClassError"""
+
+
+class ExceptionMessageError(FlexmockError):
+    """ExceptionMessageError"""
+
+
+class StateError(FlexmockError):
+    """StateError"""
+
+
+class MethodCallError(FlexmockError):
+    """MethodCallError"""
+
+
+class CallOrderError(FlexmockError):
+    """CallOrderError"""
