+++ pyproject.toml
@@ -42,6 +42,7 @@ disable = [
     "too-many-locals",
     "too-many-branches",
     "too-many-statements",
+    "too-many-public-methods",
     "protected-access",
     "broad-except",
     "fixme",
